requirements.txt
coverage
ccxt==3.0.84
eciespy
eth-brownie
enforce_typing
gql
mypy
numpy
pandas
pytest
requests
scipy
types-requests
